+++ src/Keter/HostManager.hs
@@ -24,11 +24,15 @@ import           Data.IORef
 import qualified Data.Map            as Map
 import qualified Data.Set            as Set
 import           Data.Text.Encoding  (encodeUtf8)
-import           Keter.Types
+import           Keter.Config
 import           Keter.LabelMap      (LabelMap)
 import qualified Keter.LabelMap      as LabelMap
 import           Prelude             hiding (log)
 import qualified Network.TLS as TLS
+import           Keter.Common
+import           System.FilePath            (FilePath)
+import           Data.Set                   (Set)
+import           Data.Map                   (Map)
 
 type HMState = LabelMap HostValue